requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytesseract
pillow
opencv-python
numpy
dateparser
pytz
